ml_project/requirements.txt
torch
torchvision
pandas
Pillow
pypdf
numpy